ROOT: refactored plugins structure
